requirements.txt
pyqt6-sip
# GUI Framework (will auto-install compatible versions)
PyQt6

# HTTP Client
requests

# Security & Encryption
cryptography

# Configuration Management
python-dotenv

# System Information
psutil

# Executable Builder
pyinstaller

# Logging with Colors
colorlog

# JSON Web Tokens
PyJWT